apply review feedback, fix typo
