+++ src/System.Runtime.Extensions/src/System.Runtime.Extensions.csproj
@@ -8,7 +8,6 @@
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <IsPartialFacadeAssembly>true</IsPartialFacadeAssembly>
     <IncludeDllSafeSearchPathAttribute>true</IncludeDllSafeSearchPathAttribute>
-    <DefineConstants Condition="'$(TargetGroup)' == 'uapaot' or '$(TargetGroup)' == 'uap'">$(DefineConstants);uap</DefineConstants>
   </PropertyGroup>
   <!-- Default configurations to help VS understand the configurations -->
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'netcoreapp-Unix-Debug|AnyCPU'" />
@@ -92,7 +91,7 @@
   <!-- WINDOWS: Shared CoreCLR and .NET Native -->
   <ItemGroup Condition="'$(TargetsWindows)' == 'true'">
     <Compile Include="System\Environment.Windows.cs" />
-    <Compile Include="System\Environment.Win32.cs" />
+    <Compile Include="System\Environment.Win32.cs"  Condition="'$(TargetGroup)' != 'uapaot' and '$(TargetGroup)' != 'uap'" />
     <Compile Include="System\Runtime\Versioning\VersioningHelper.Windows.cs" />
     <Compile Include="System\Diagnostics\Stopwatch.Windows.cs" />
     <Compile Include="$(CommonPath)\System\IO\Win32Marshal.cs">
